move  to MeshIO. import load_mesh into base discretize namespace. deserialize onto BaseMesh rather than properties.HasProperties
