+++ discretize/utils/__init__.py
@@ -2,14 +2,14 @@ from __future__ import print_function
 
 from .matutils import (
     mkvc, sdiag, sdInv, speye, kron3, spzeros, ddx, av,
-   av_extrap, ndgrid, ind2sub, sub2ind, getSubArray,
-   inv3X3BlockDiagonal, inv2X2BlockDiagonal, TensorType,
-   makePropertyTensor, invPropertyTensor, Zero,
-   Identity
+    av_extrap, ndgrid, ind2sub, sub2ind, getSubArray,
+    inv3X3BlockDiagonal, inv2X2BlockDiagonal, TensorType,
+    makePropertyTensor, invPropertyTensor, Zero,
+    Identity
 )
 from .codeutils import (isScalar, asArray_N_x_Dim)
 from .meshutils import (
-    load_mesh, exampleLrmGrid, meshTensor, closestPoints, ExtractCoreMesh,
+    exampleLrmGrid, meshTensor, closestPoints, ExtractCoreMesh,
     random_model
 )
 from .curvutils import volTetra, faceInfo, indexCube
